+++ nginx_ldap_auth/__init__.py
@@ -1,4 +1,4 @@
-__version__: str = "2.5.1"
+__version__: str = "2.6.1"
 __author__: str = "Caltech IMSS ADS"
 __author_email__: str = "imss-ads-staff@caltech.edu"
 __description__: str = (